Move the forced theme variant to internal test and use
