--- cmd/fyne_demo/theme.go
@@ -1,17 +1,0 @@
-package main
-
-import (
-	"image/color"
-
-	"fyne.io/fyne/v2"
-)
-
-type forcedVariant struct {
-	fyne.Theme
-
-	variant fyne.ThemeVariant
-}
-
-func (f *forcedVariant) Color(name fyne.ThemeColorName, _ fyne.ThemeVariant) color.Color {
-	return f.Theme.Color(name, f.variant)
-}
